requirements.txt
pyautogui>=0.9.54
pynput>=1.7.6
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
pygetwindow>=0.0.9
pydantic>=2.0
